moved stuff + removed model_index
